+++ bigtangle-server/src/test/java/net/bigtangle/core/SerializationTest.java
@@ -9,6 +9,7 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 
 import java.io.IOException;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
@@ -44,7 +45,7 @@ public class SerializationTest {
     @Test
     public void testTokenInfoSerialization() throws JsonParseException, JsonMappingException, IOException {
         List<MultiSignAddress> addresses = new ArrayList<>();
-        Token tokens = Token.buildSimpleTokenInfo(true, "1", "2", "3", "4", 2, 3, 4, true, 0, "de");
+        Token tokens = Token.buildSimpleTokenInfo(true, "1", "2", "3", "4", 2, 3, BigInteger.valueOf(4), true, 0, "de");
         TokenInfo info1 = new TokenInfo();
         info1.setToken(tokens);
         info1.setMultiSignAddresses(addresses);
