+++ bigtangle-server/src/test/java/net/bigtangle/server/SubtangleIntegrationTests.java
@@ -157,6 +157,6 @@ public class SubtangleIntegrationTests extends AbstractIntegrationTest {
         Coin coinbase = getBalanceCoin(subtangleKey, NetworkParameters.BIGTANGLE_TOKENID);
         logger.info("get balance coin : " + coinbase);
 
-        assertTrue(amount == coinbase.getValue());
+        assertTrue(amount == coinbase.getValue().longValue());
     }
 }
